refactor(crypto): extract shared canonical hash from approval modules
Move canonicalHash() and sortKeysDeep() into src/crypto/canonical-hash.ts,
eliminating the "keep in sync" duplication between google-services/approval.ts
and relay/approval-token.ts. Both consumers now import from the shared module.
